Changed for any distribution that can git clone
Functional enumeration scripts. Changed for any distribution that has E4L & dirb - you may have to edit it yourself. This was a quick build to run through boxes in the OSCP as host to host to host manual scanning is very inefficient. This also allows you to categorize hosts and keep that information for later without the extra effort of command > stdout.txt for each and every scan. It's simple, but effective.